fix: github_trending example removed - it generates more issues (because of scraping tool) than was worth it
